requirements.txt
openpyxl==3.1.2
pandas==2.0.3
streamlit==1.27.2
xlrd==2.0.1
msoffcrypto-tool==5.1.1